backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
mangum==0.17.0
requests==2.32.3
python-jose==3.3.0
python-dotenv==1.0.1
